Bethe Free Energy implementation in RxInfer

RxInfer.AbstractScoreObjective — Type

AbstractScoreObjective

Abstract type for functional objectives that can be used in the score function.

Note

score is defined in the ReactiveMP package.

RxInfer.BetheFreeEnergy — Type

BetheFreeEnergy(marginal_skip_strategy, scheduler, diagnostic_checks)

Creates Bethe Free Energy values stream when passed to the score function.

RxInfer.apply_diagnostic_check — Function

apply_diagnostic_check(check, context, stream)

This function applies a check to the stream. Accepts optional context object for custom error messages.

RxInfer.BetheFreeEnergyCheckNaNs — Type

BetheFreeEnergyCheckNaNs

If enabled checks that both variable and factor bound score functions in Bethe Free Energy computation do not return NaNs. Throws an error if finds NaN.

See also: BetheFreeEnergyCheckInfs

RxInfer.BetheFreeEnergyCheckInfs — Type

BetheFreeEnergyCheckInfs

If enabled checks that both variable and factor bound score functions in Bethe Free Energy computation do not return Infs. Throws an error if finds Inf.

See also: BetheFreeEnergyCheckNaNs