Inference results postprocessing

Both inference and rxinference allow users to postprocess the inference result with the postprocess = ... keyword argument. The inference engine operates on wrapper types to distinguish between marginals and messages. By default these wrapper types are removed from the inference results if no addons option is present. Together with the enabled addons, however, the wrapper types are preserved in the inference result output value. Use the options below to change this behaviour:

RxInfer.DefaultPostprocess — Type

DefaultPostprocess picks the most suitable postprocessing step automatically

RxInfer.UnpackMarginalPostprocess — Type

This postprocessing step removes the Marginal wrapper type from the result

RxInfer.NoopPostprocess — Type

This postprocessing step does nothing